@@ -1,7 +1,5 @@
 import 'dart:async';
-import 'dart:io';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:pangolin/pangolin.dart';
 
@@ -21,14 +19,14 @@ Stream<BasePangolinResponse> get pangolinResponseEventHandler =>
     _pangolinResponseEventHandlerController.stream;
 
 Future<bool> registerPangolin({
-  @required String appId,
-  @required bool useTextureView,
-  @required String appName,
-  @required bool allowShowNotify,
-  @required bool allowShowPageWhenScreenLock,
-  @required bool debug,
-  @required bool supportMultiProcess,
-  List<int> directDownloadNetworkType,
+  required String appId,
+  required bool useTextureView,
+  required String appName,
+  required bool allowShowNotify,
+  required bool allowShowPageWhenScreenLock,
+  required bool debug,
+  required bool supportMultiProcess,
+  List<int>? directDownloadNetworkType,
 }) async {
   return await _channel.invokeMethod("register", {
     "appId": appId,
@@ -49,23 +47,23 @@ Future<bool> registerPangolin({
 }
 
 Future<bool> loadSplashAd(
-    {@required String mCodeId, @required bool debug}) async {
+    {required String mCodeId, required bool debug}) async {
   return await _channel
       .invokeMethod("loadSplashAd", {"mCodeId": mCodeId, "debug": debug});
 }
 
 Future loadRewardAd({
-  @required String mCodeId,
-  @required bool debug,
-  @required bool supportDeepLink,
-  @required String rewardName,
-  @required int rewardAmount,
-  @required bool isExpress,
-  double expressViewAcceptedSizeH,
-  double expressViewAcceptedSizeW,
-  @required userID,
-  String mediaExtra,
-  @required bool isHorizontal,
+  required String mCodeId,
+  required bool debug,
+  required bool supportDeepLink,
+  required String rewardName,
+  required int rewardAmount,
+  required bool isExpress,
+  double? expressViewAcceptedSizeH,
+  double? expressViewAcceptedSizeW,
+  required userID,
+  String? mediaExtra,
+  required bool isHorizontal,
 }) async {
   return await _channel.invokeMethod("loadRewardAd", {
     "mCodeId": mCodeId,
@@ -82,15 +80,14 @@ Future loadRewardAd({
   });
 }
 
-Future loadBannerAd({
-  @required String mCodeId,
-  @required bool supportDeepLink,
-  double expressViewWidth,
-  double expressViewHeight,
-  bool isCarousel,
-  int interval,
-  int topMargin
-}) async {
+Future loadBannerAd(
+    {required String mCodeId,
+    required bool supportDeepLink,
+    double? expressViewWidth,
+    double? expressViewHeight,
+    bool? isCarousel,
+    int? interval,
+    int? topMargin}) async {
   return await _channel.invokeMethod("loadBannerAd", {
     "mCodeId": mCodeId,
     "supportDeepLink": supportDeepLink,
@@ -102,20 +99,18 @@ Future loadBannerAd({
   });
 }
 
-Future loadInterstitialAd({
-  @required String mCodeId,
-  double expressViewWidth,
-  double expressViewHeight
-}) async {
-  return await _channel.invokeMethod("loadInterstitialAd",{
+Future loadInterstitialAd(
+    {required String mCodeId,
+    double? expressViewWidth,
+    double? expressViewHeight}) async {
+  return await _channel.invokeMethod("loadInterstitialAd", {
     "mCodeId": mCodeId,
     "expressViewWidth": expressViewWidth,
     "expressViewHeight": expressViewHeight,
   });
 }
 
-Future removeBannerAd() async
-{
+Future removeBannerAd() async {
   await _channel.invokeMethod("removeBannerAd");
 }
 
